feat: move entity to db/entities
