+++ src/db/entities/sendgrid-email.entity.ts
